feat(watch): transcript output formats (txt/srt/vtt/md) + extract dialog mixin
WATCH-9 follow-up. The offline "Transcribe audio (Whisperer)" watch action can
now write the transcript as plain Text, SubRip (.srt), WebVTT (.vtt), or
Markdown, chosen per profile:

- WhispererTranscribeAction's transcriber now returns the full TranscriptionResult
  (not just text), and run() renders the chosen format via the existing
  speech/formatters.py, writing the sibling file with the matching extension.
- SRT/VTT need timestamped segments and fall back to plain text when the engine
  returns none, so a profile never writes an empty caption file. Unknown formats
  are flagged in validate().
- A "Transcript format" picker was added to the Add/Edit Watch Profile dialog.

To stay under the GATE-11 budget (main_frame.py was at its ceiling), the dialog's
_edit_watch_profile (430 lines) was extracted into a new
quill/ui/main_frame_watch_profile.py (WatchProfileDialogMixin); the picker rows
moved with it. main_frame.py budget ratcheted 25499 -> 25093. Dialog inventory
re-registered the two surfaces under the new path (same classifications).

Tests: new format cases (srt/vtt/md, no-segment fallback, unknown-format
validation) in test_watch_actions.py. Docs: CHANGELOG + userguide (+ artifacts).
