+++ CHANGELOG.md
@@ -17,6 +17,7 @@ First pre-release of the 0.7.0 line. Tagged off `release/0.7.0-verbosity` once t
 - **Offline speech lives under Tools > Speech > Whisperer (#617).** The on-device transcription, captions, dictation, and voice-command items moved out of the AI menu — they need no AI account or key, so they no longer imply an AI dependency. The Windows installer's optional `speechwhisper` component (whisper.cpp engine payload under `tools\speech\whispercpp`) is now emitted by the installer-script generator so a build keeps the offline engine.
 - **Hugging Face download hardening + transparency (#617).** Three changes to how speech models are fetched: (1) the model manager shows each model's license and a model-card URL in the row label and the download confirm (`service.model_card_url` derives it from either the whisper.cpp resolve URL or the FW repo id); (2) supply-chain pinning — whisper.cpp models download from a pinned commit with a per-file `sha256` (the existing verifier now runs), Faster Whisper repos pin a `revision` (new `SpeechModelInfo.revision`), and the stale tinydiarize speaker-detection URL (which 404'd against `ggerganov/whisper.cpp`) was fixed to `akashmjn/tinydiarize-whisper.cpp`; (3) an optional Hugging Face token (new wx-free `quill/core/speech/hf_auth.py`, reusing the `QUILL:huggingface:token` credential) is attached to both download paths when present, a 429 becomes a clear "add a token or retry" message, and `Tools > Speech > Whisperer > Hugging Face Token...` sets it via a masked dialog. A token is not required for the public MIT models. Tests: `test_hf_auth.py`, plus catalog-pinning and license/card-url assertions.
 - **Watch Folder can now auto-transcribe arriving audio on-device (WATCH-9).** A new **Transcribe audio (Whisperer)** watch action runs whisper.cpp / Faster Whisper entirely on your machine over each audio or video file a watch profile claims, writing a sibling `.txt` transcript — nothing is uploaded, so it needs no consent (the on-device counterpart to the existing cloud `Transcribe audio (OpenAI Whisper)` action). It replaces the old "not available yet" placeholder, is discoverable in the watch action picker, and reports clear guidance (open Tools > Speech > Whisperer > Manage Speech Models) when no offline model is installed rather than failing silently. New wx-free `quill/core/speech/transcribe.py` picks the best available offline provider and model; the two transcription watch actions moved into a new `quill/core/watch_transcribe.py` to keep `watch_actions.py` within its size budget. Tests: `test_transcribe.py` (8) and the WhispererTranscribeAction cases in `test_watch_actions.py`. This is the first item delivered from the BITS Whisperer consolidation workstream.
+- **Choose the transcript output format: Text, SubRip (.srt), WebVTT (.vtt), or Markdown (WATCH-9).** The offline transcribe watch action now honors an `output_format` option and writes the sibling file with the matching extension, reusing `speech/formatters.py` (the action returns the full timestamped result, not just text). SRT/VTT carry timestamps and fall back to plain text when the engine returns no segments, so a profile never writes an empty caption file. A **Transcript format** picker was added to the Add/Edit Watch Profile dialog. To stay under the GATE-11 size budget, that dialog's `_edit_watch_profile` was extracted from `main_frame.py` into a new `quill/ui/main_frame_watch_profile.py` (`WatchProfileDialogMixin`), ratcheting the `main_frame.py` budget down by ~430 lines.
 
 ### Text-file round-trip and accessibility fixes
