+++ docs/user guide/userguide.html
@@ -1832,8 +1832,14 @@ opening, moving, copying, converting, running a macro or a sandboxed
 transform, and OCR, a profile can <strong>Transcribe audio
 (Whisperer)</strong>: any audio or video file dropped into the folder is
 transcribed on your machine with the offline Whisperer engine and a
-<code>.txt</code> transcript is written next to it — nothing is
-uploaded. It needs no consent; if no speech model is installed yet, the
+transcript is written next to it — nothing is uploaded. Choose the
+<strong>Transcript format</strong> in the profile: plain
+<strong>Text</strong> (<code>.txt</code>), <strong>SubRip
+captions</strong> (<code>.srt</code>), <strong>WebVTT captions</strong>
+(<code>.vtt</code>), or <strong>Markdown</strong> (<code>.md</code>).
+The caption formats carry timestamps; if the engine returns no timed
+segments they fall back to plain text so you never get an empty caption
+file. It needs no consent; if no speech model is installed yet, the
 profile tells you to download one from <strong>Tools → Speech →
 Whisperer → Manage Speech Models</strong>. (A separate
 <strong>Transcribe audio (OpenAI Whisper)</strong> action is available
