+++ quill/tools/module_size_budgets.json
@@ -72,7 +72,7 @@
   "budgets": {
     "quill/core/format_ops.py": 617,
     "quill/ui/main_frame_speech.py": 836,
-    "quill/ui/main_frame.py": 25499,
+    "quill/ui/main_frame.py": 25093,
     "quill/ui/main_frame_quill_key.py": 753,
     "quill/ui/main_frame_braille.py": 775,
     "quill/ui/assistant_tools.py": 2008,
@@ -227,5 +227,6 @@
   "_rebaseline_2026_06_20_615_data_location_wizard": "setup_wizard_pages.py 1080->1201 (+121) for #615 follow-up: new _DataLocationPage (~115 lines) lets the first-run wizard choose AppData (default), portable (only offered when storage_mode.portable_root_dir() resolves to a verified bundle), or a custom folder via wx.DirDialog. Wired into _build_all_pages()/_rebuild_active() as the new page 5 (Summary moves to page 6); _SummaryPage.update_summary() gained a 3-line addition to surface the choice. Also replaced the now-redundant, drifted _PAGE_TITLES position-indexed list with page.GetName() in _show_page's announcement (each page already sets a descriptive name via _WizardPage.__init__) -- a net code-size wash that incidentally fixes a pre-existing announcement bug for the AI Provider/Keyboard/Summary pages whenever the conditional AI page is skipped. CQ-1 decomposition is still the path to lower the budget.",
   "_rebaseline_2026_06_20_620_simple_open_dialog": "#620 Simple File Open dialog: settings.py +1 (use_simple_file_dialog bool, default False) and settings_specs.py +14 (the matching SettingSpec in the general group) for the new opt-in file picker. main_frame.py 25101->25092 (-9) for replacing the inline wx.FileDialog block in open_file() with a single dispatcher call -- the wx.FileDialog construction itself (and the long wildcard list) moved into the new SimpleOpenMixin._prompt_native_open_dialog so the native dialog still appears when the setting is off. New files: quill/ui/simple_open_dialog.py (494 lines, given explicit 500 budget entry -- over the 600 default cap would otherwise cause GATE-11 to fail when the dialog grows with new features like column sorting or breadcrumb navigation) and quill/ui/main_frame_simple_open.py (180 lines, given explicit 200 budget entry). Both are tracked from day one so the next grow triggers an explicit re-baseline. settings_specs.py also grew by 5 more lines from the ruff format pass on the new SettingSpec (multi-line strings were reformatted), tracked together.",
   "_rebaseline_2026_06_20_620_settings": "settings.py 908->917 (+9) for #620: new use_simple_file_dialog bool field (default False) plus the from_dict loader line and constructor kwarg. Tracked alongside the matching settings_specs.py rebaseline.",
-  "_rebaseline_2026_06_20_606_theme_guard": "main_frame.py 25092->25101 (+9) for the #606 follow-up crash: on a fresh install, _apply_theme was called from __init__ before _create_document_tab ran (the wizard-deferral path), so self.editor was unset and the three editor.* calls raised AttributeError, closing QUILL with an unexpected-error dialog. _apply_theme now guards the three editor writes with getattr(self, 'editor', None) and skips the editor.Refresh() when the editor does not yet exist, mirroring the getattr() pattern already used in show() and _focus_editor(). The chrome + statusbar writes always apply, so the wizard still sees the correct chrome even before the first document tab. The +9 is the getattr() local + 7-line comment block explaining the wizard-deferral interaction."
+  "_rebaseline_2026_06_20_606_theme_guard": "main_frame.py 25092->25101 (+9) for the #606 follow-up crash: on a fresh install, _apply_theme was called from __init__ before _create_document_tab ran (the wizard-deferral path), so self.editor was unset and the three editor.* calls raised AttributeError, closing QUILL with an unexpected-error dialog. _apply_theme now guards the three editor writes with getattr(self, 'editor', None) and skips the editor.Refresh() when the editor does not yet exist, mirroring the getattr() pattern already used in show() and _focus_editor(). The chrome + statusbar writes always apply, so the wizard still sees the correct chrome even before the first document tab. The +9 is the getattr() local + 7-line comment block explaining the wizard-deferral interaction.",
+  "_rebaseline_2026_06_22_watch9_transcript_format": "WATCH-9 transcript-format picker: the Add/Edit Watch Profile dialog gained a Transcript format choice (Text/SRT/VTT/Markdown) for the offline Transcribe action. main_frame.py was at its 25499 ceiling, so _edit_watch_profile (430 lines) was extracted to a new quill/ui/main_frame_watch_profile.py (WatchProfileDialogMixin), dropping main_frame.py 25528->25093 (extraction plus 8 now-unused imports removed). Budget ratcheted down 25499->25093. New mixin file is 454 lines (under the 600 default cap, no explicit entry). The picker rows moved with the method into the mixin."
 }
