+++ tests/unit/ui/fixtures/dialog_inventory.json
@@ -56,8 +56,6 @@
   "quill/ui/main_frame.py::MainFrame._choose_searchable_option::wx.Dialog": "hardened_custom",
   "quill/ui/main_frame.py::MainFrame._confirm_beta_channel::wx.MessageDialog": "native",
   "quill/ui/main_frame.py::MainFrame._confirm_restart_for_data_location::wx.MessageDialog": "native",
-  "quill/ui/main_frame.py::MainFrame._edit_watch_profile._pick_folder::wx.DirDialog": "native",
-  "quill/ui/main_frame.py::MainFrame._edit_watch_profile::wx.Dialog": "hardened_custom",
   "quill/ui/main_frame.py::MainFrame._html_info::wx.MessageDialog": "native",
   "quill/ui/main_frame.py::MainFrame._maybe_prompt_braille_pack_install::wx.MessageDialog": "native",
   "quill/ui/main_frame.py::MainFrame._offer_beta_switch::wx.MessageDialog": "native",
@@ -256,6 +254,8 @@
   "quill/ui/main_frame_speech.py::SpeechCommandsMixin.transcribe_audio_offline::wx.FileDialog": "native",
   "quill/ui/main_frame_ssh.py::SshEditingMixin._request_for_site::wx.PasswordEntryDialog": "native",
   "quill/ui/main_frame_verbosity.py::VerbosityCommandsMixin.open_verbosity_preferences::wx.Dialog": "hardened_custom",
+  "quill/ui/main_frame_watch_profile.py::WatchProfileDialogMixin._edit_watch_profile._pick_folder::wx.DirDialog": "native",
+  "quill/ui/main_frame_watch_profile.py::WatchProfileDialogMixin._edit_watch_profile::wx.Dialog": "hardened_custom",
   "quill/ui/ocr_review_dialog.py::OcrReviewDialog.__init__::wx.Dialog": "hardened_custom",
   "quill/ui/palette.py::CommandPaletteDialog.__init__::wx.Dialog": "hardened_custom",
   "quill/ui/palette.py::GoToAnythingDialog.__init__::wx.Dialog": "hardened_custom",
